+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Grizzled Scala Change Log
 
+Version 2.4.1:
+
+* Various fixes for Windows, including:
+    - Fixed `grizzled.string.WordWrapper` to use the `line.separator`
+      property, rather than a hard-coded "\n", to wrap lines.
+
 Version 2.4.0:
 
 * Removed use of `asInstanceOf` in `grizzled.io.Implicits.RichReader` and